# laravel-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel-api documentation",
        "contact": {
            "email": "zaq22022015@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1",
            "description": "Laravel-api server"
        }
    ],
    "paths": {
        "/notebook": {
            "get": {
                "tags": [
                    "Notebook"
                ],
                "summary": "Get all notes",
                "operationId": "NotesAll",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "The page number",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "tags": [
                    "Notebook"
                ],
                "summary": "Adding note in db",
                "operationId": "AddNote",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "fio": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "company": {
                                        "type": "string"
                                    },
                                    "birthday": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "fio": "aa vv bb",
                                    "email": "awdawd@a.aw",
                                    "phone": "12345678",
                                    "company": "mvd",
                                    "birthday": "1985-07-31"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    }
                }
            }
        },
        "/notebook/{id}": {
            "get": {
                "tags": [
                    "Notebook"
                ],
                "summary": "Get note by id",
                "operationId": "GetNoteById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The id of nore",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Note is found"
                    },
                    "404": {
                        "description": "Note is not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Notebook"
                ],
                "summary": "Update note by id",
                "operationId": "UpdateNoteById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The id of note",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "fio": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "company": {
                                        "type": "string"
                                    },
                                    "birthday": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "fio": "aa vv bb",
                                    "email": "awdawd@a.aw",
                                    "phone": "12345678",
                                    "company": "mvd",
                                    "birthday": "1985-07-31"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Updated"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Notebook"
                ],
                "summary": "Delete note by id",
                "operationId": "DeleteNoteById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The id of note",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Note is deleted"
                    },
                    "404": {
                        "description": "Note is not found"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Notebook",
            "description": "Work with notes"
        }
    ]
}